refactor: splitting up the large files into modules
