+++ src/da/mock.rs
@@ -1,95 +1,14 @@
-use crate::{
-    consts::CHANNEL_BUFFER_SIZE,
-    error::{DAResult, DataAvailabilityError, DeimosResult, GeneralError},
-    utils::Signable,
-    zk_snark::{Bls12Proof, VerifyingKey},
-};
+use crate::error::{DAResult, DataAvailabilityError};
 use async_trait::async_trait;
-use celestia_rpc::{BlobClient, Client, HeaderClient};
-use celestia_types::{blob::GasPrice, nmt::Namespace, Blob};
-use ed25519::Signature;
 use fs2::FileExt;
-use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use std::{
     self,
     fs::{File, OpenOptions},
     io::{Read, Seek, Write},
-    str::FromStr,
-    sync::Arc,
-};
-use tokio::{
-    sync::{
-        mpsc::{channel, Receiver, Sender},
-        Mutex,
-    },
-    task::spawn,
 };
 
-#[derive(Serialize, Deserialize, Clone)]
-pub struct EpochJson {
-    pub height: u64,
-    pub prev_commitment: String,
-    pub current_commitment: String,
-    pub proof: Bls12Proof,
-    pub verifying_key: VerifyingKey,
-    pub signature: Option<String>,
-}
-
-impl TryFrom<&Blob> for EpochJson {
-    type Error = GeneralError;
-
-    fn try_from(value: &Blob) -> Result<Self, GeneralError> {
-        // convert blob data to utf8 string
-        let data_str = String::from_utf8(value.data.clone()).map_err(|e| {
-            GeneralError::EncodingError(format!("encoding blob data to utf8 string: {}", e))
-        })?;
-
-        serde_json::from_str(&data_str)
-            .map_err(|e| GeneralError::DecodingError(format!("epoch json: {}", e)))
-    }
-}
-
-impl Signable for EpochJson {
-    fn get_signature(&self) -> DeimosResult<Signature> {
-        match &self.signature {
-            Some(signature) => Signature::from_str(signature)
-                .map_err(|e| GeneralError::ParsingError(format!("signature: {}", e)).into()),
-            None => Err(GeneralError::MissingArgumentError("signature".to_string()).into()),
-        }
-    }
-
-    fn get_content_to_sign(&self) -> DeimosResult<String> {
-        let mut copy = self.clone();
-        copy.signature = None;
-        serde_json::to_string(&copy).map_err(|e| GeneralError::EncodingError(e.to_string()).into())
-    }
-
-    fn get_public_key(&self) -> DeimosResult<String> {
-        //TODO(@distractedm1nd): the below comment isn't good enough of an argument to not return the public key, it should be fixed
-
-        // for epoch json the public key to verify is the one from the sequencer which should be already be public and known from every light client
-        // so if we use this function there should be an error
-        Err(GeneralError::MissingArgumentError("public key".to_string()).into())
-    }
-}
-
-#[async_trait]
-pub trait DataAvailabilityLayer: Send + Sync {
-    async fn get_message(&self) -> DAResult<u64>;
-    async fn initialize_sync_target(&self) -> DAResult<u64>;
-    async fn get(&self, height: u64) -> DAResult<Vec<EpochJson>>;
-    async fn submit(&self, epoch: &EpochJson) -> DAResult<u64>;
-    async fn start(&self) -> DAResult<()>;
-}
-
-pub struct CelestiaConnection {
-    pub client: celestia_rpc::Client,
-    pub namespace_id: Namespace,
-
-    synctarget_tx: Arc<Sender<u64>>,
-    synctarget_rx: Arc<Mutex<Receiver<u64>>>,
-}
+use crate::da::{DataAvailabilityLayer, EpochJson};
 
 /// The `NoopDataAvailabilityLayer` is a mock implementation of the `DataAvailabilityLayer` trait.
 pub struct NoopDataAvailabilityLayer {}
@@ -123,173 +42,6 @@ impl DataAvailabilityLayer for NoopDataAvailabilityLayer {
 ///
 /// This implementation is intended for testing and development only and should not be used in production environments. It provides a way to test the interactions with the data availability layer without the overhead of real network communication or data persistence.
 pub struct LocalDataAvailabilityLayer {}
-
-impl CelestiaConnection {
-    // TODO: Should take config
-    pub async fn new(
-        connection_string: &String,
-        auth_token: Option<&str>,
-        namespace_hex: &String,
-    ) -> DAResult<Self> {
-        let (tx, rx) = channel(CHANNEL_BUFFER_SIZE);
-
-        let client = Client::new(&connection_string, auth_token)
-            .await
-            .map_err(|e| {
-                DataAvailabilityError::ConnectionError(format!(
-                    "websocket initialization failed: {}",
-                    e
-                ))
-            })?;
-
-        let decoded_hex = match hex::decode(namespace_hex) {
-            Ok(hex) => hex,
-            Err(e) => {
-                return Err(DataAvailabilityError::GeneralError(
-                    GeneralError::DecodingError(format!(
-                        "decoding namespace '{}': {}",
-                        namespace_hex, e
-                    )),
-                ))
-            }
-        };
-
-        let namespace_id = Namespace::new_v0(&decoded_hex).map_err(|e| {
-            DataAvailabilityError::GeneralError(GeneralError::EncodingError(format!(
-                "creating namespace '{}': {}",
-                namespace_hex, e
-            )))
-        })?;
-
-        Ok(CelestiaConnection {
-            client,
-            namespace_id,
-            synctarget_tx: Arc::new(tx),
-            synctarget_rx: Arc::new(Mutex::new(rx)),
-        })
-    }
-}
-
-#[async_trait]
-impl DataAvailabilityLayer for CelestiaConnection {
-    async fn get_message(&self) -> DAResult<u64> {
-        match self.synctarget_rx.lock().await.recv().await {
-            Some(height) => Ok(height),
-            None => Err(DataAvailabilityError::ChannelReceiveError),
-        }
-    }
-
-    async fn initialize_sync_target(&self) -> DAResult<u64> {
-        match HeaderClient::header_network_head(&self.client).await {
-            Ok(extended_header) => Ok(extended_header.header.height.value()),
-            Err(err) => Err(DataAvailabilityError::NetworkError(format!(
-                "getting network head from da layer: {}",
-                err
-            ))),
-        }
-    }
-
-    async fn get(&self, height: u64) -> DAResult<Vec<EpochJson>> {
-        trace!("searching for epoch on da layer at height {}", height);
-        match BlobClient::blob_get_all(&self.client, height, &[self.namespace_id]).await {
-            Ok(blobs) => {
-                let mut epochs = Vec::new();
-                for blob in blobs.iter() {
-                    match EpochJson::try_from(blob) {
-                        Ok(epoch_json) => epochs.push(epoch_json),
-                        Err(_) => {
-                            DataAvailabilityError::GeneralError(GeneralError::ParsingError(
-                                format!(
-                                    "marshalling blob from height {} to epoch json: {}",
-                                    height,
-                                    serde_json::to_string(&blob).unwrap()
-                                ),
-                            ));
-                        }
-                    }
-                }
-                Ok(epochs)
-            }
-            Err(err) => {
-                // todo: this is a hack to handle a retarded error from cel-node that will be fixed in v0.15.0
-                if err.to_string().contains("blob: not found") {
-                    Ok(vec![])
-                } else {
-                    Err(DataAvailabilityError::DataRetrievalError(
-                        height,
-                        format!("getting epoch from da layer: {}", err),
-                    ))
-                }
-            }
-        }
-    }
-
-    async fn submit(&self, epoch: &EpochJson) -> DAResult<u64> {
-        debug!("posting epoch {} to da layer", epoch.height);
-
-        let data = serde_json::to_string(&epoch).map_err(|e| {
-            DataAvailabilityError::GeneralError(GeneralError::ParsingError(format!(
-                "serializing epoch json: {}",
-                e
-            )))
-        })?;
-        let blob = Blob::new(self.namespace_id.clone(), data.into_bytes()).map_err(|e| {
-            DataAvailabilityError::GeneralError(GeneralError::BlobCreationError(e.to_string()))
-        })?;
-        trace!("blob: {:?}", serde_json::to_string(&blob).unwrap());
-        match self
-            .client
-            .blob_submit(&[blob.clone()], GasPrice::from(-1.0))
-            .await
-        {
-            Ok(height) => Ok(height),
-            Err(err) => Err(DataAvailabilityError::SubmissionError(
-                epoch.height,
-                err.to_string(),
-            )),
-        }
-    }
-
-    async fn start(&self) -> DAResult<()> {
-        let mut header_sub = HeaderClient::header_subscribe(&self.client)
-            .await
-            .map_err(|e| {
-                DataAvailabilityError::NetworkError(format!(
-                    "subscribing to headers from da layer: {}",
-                    e
-                ))
-            })?;
-
-        let synctarget_buffer = self.synctarget_tx.clone();
-        spawn(async move {
-            while let Some(extended_header_result) = header_sub.next().await {
-                match extended_header_result {
-                    Ok(extended_header) => {
-                        let height = extended_header.header.height.value();
-                        match synctarget_buffer.send(height).await {
-                            Ok(_) => {
-                                debug!("sent sync target update for height {}", height);
-                            }
-                            Err(_) => {
-                                DataAvailabilityError::SyncTargetError(format!(
-                                    "sending sync target update message for height {}",
-                                    height
-                                ));
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        DataAvailabilityError::NetworkError(format!(
-                            "retrieving header from da layer: {}",
-                            e
-                        ));
-                    }
-                }
-            }
-        });
-        Ok(())
-    }
-}
 
 impl LocalDataAvailabilityLayer {
     pub fn new() -> Self {
@@ -379,12 +131,12 @@ impl DataAvailabilityLayer for LocalDataAvailabilityLayer {
 }
 
 #[cfg(test)]
-mod da_tests {
+mod tests {
     use crate::{
         utils::validate_epoch,
         zk_snark::{
             deserialize_custom_to_verifying_key, deserialize_proof, serialize_proof,
-            serialize_verifying_key_to_custom, BatchMerkleProofCircuit, VerifyingKey,
+            serialize_verifying_key_to_custom, BatchMerkleProofCircuit, Bls12Proof, VerifyingKey,
         },
     };
